2023.02.24

+++ Client/CPathMgr.cpp
@@ -19,7 +19,7 @@ void CPathMgr::init()
 	GetCurrentDirectory(255, m_szContentPath);
 
 	// 상위 폴더로 나가기
-	int iLen = wcslen(m_szContentPath); // 전체 길이를 알아서 뒤에서부터 \\ 가 있는 곳을 찾아서 끊어두자
+	int iLen = (int)wcslen(m_szContentPath); // 전체 길이를 알아서 뒤에서부터 \\ 가 있는 곳을 찾아서 끊어두자
 	for (int i = iLen - 1; 0 <= i; --i)
 	{
 		if ('\\' == m_szContentPath[i])
